+++ code/src/components/Icons.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const Icons = () => {
+  return (
+    <div className="icons-container">
+      <img src="./icons/heart.svg" className="heart-icon" alt="heart icon" />
+      <img src="./icons/play.svg" className="play-icon" alt="play icon" />
+      <img src="./icons/dots.svg" className="dots-icon" alt="dots icon" />
+    </div>
+  )
+}
