+++ code/src/components/Title.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const Title = (props) => {
+  return (
+    <div className="album-name">
+      <h1 className="album-name">{props.title}</h1>
+    </div>
+  )
+}
